Reorganised file structure to have separate folders for pages, components and styles
