+++ src/pages/ContactUs.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ContactUs = () => {
+  return (
+    <div>ContactUs</div>
+  )
+}
+
+export default ContactUs
