+++ src/pages/Pricing.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Pricing = () => {
+  return (
+    <div>Pricing</div>
+  )
+}
+
+export default Pricing
