+++ src/pages/Products.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Products = () => {
+  return (
+    <div>Products</div>
+  )
+}
+
+export default Products
